Rename GacelaConfigArgs dir to GacelaConfigBuilder
