+++ tests/Feature/Framework/UsingConfigTypePhp/gacela.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 use Gacela\Framework\AbstractConfigGacela;
 use Gacela\Framework\Config\ConfigReader\PhpConfigReader;
-use Gacela\Framework\Config\GacelaConfigArgs\ConfigBuilder;
+use Gacela\Framework\Config\GacelaConfigBuilder\ConfigBuilder;
 
 return static fn () => new class () extends AbstractConfigGacela {
     public function config(ConfigBuilder $configBuilder): void
