+++ tests/Feature/Framework/UsingConfigWithBootstrapSetup/FeatureTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace GacelaTest\Feature\Framework\UsingConfigWithBootstrapSetup;
 
 use Gacela\Framework\Config\ConfigReader\PhpConfigReader;
-use Gacela\Framework\Config\GacelaConfigArgs\ConfigBuilder;
+use Gacela\Framework\Config\GacelaConfigBuilder\ConfigBuilder;
 use Gacela\Framework\Gacela;
 use PHPUnit\Framework\TestCase;
